requirements.txt
humanfriendly==10.0
prometheus_client==0.21.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
